requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
sentence-transformers
Pillow
python-dotenv